Programming! City editing

+++ app/src/main/java/com/example/caxidy/tourishare/Ciudad.java
@@ -53,7 +53,7 @@ public class Ciudad {
         return latitud;
     }
 
-    public void setLatitud(float latitud) {
+    public void setLatitud(double latitud) {
         this.latitud = latitud;
     }
 
@@ -61,7 +61,7 @@ public class Ciudad {
         return longitud;
     }
 
-    public void setLongitud(float longitud) {
+    public void setLongitud(double longitud) {
         this.longitud = longitud;
     }
